[Operators] Add dilations member to ConvolutionNode
Description:
This commit adds a vector member named dilations to ConvolutionNode in
preparation for adding support for dilated convolutions. It also updates
the interface and implementation of glow::Function::createConv to accept
a dilation(s) parameter, repeat it in an array if applicable, and pass
it to the ConvolutionNode. Finally, the commit updates all uses of
glow::Function::createConv to pass in a dilation(s) parameter.

Testing:
All unit tests pass.